Version 3.1.0

+++ library/src/main/java/com/appdimens/dynamic/code/auto/DimenAutoScaled.kt
@@ -290,7 +290,7 @@ class DimenAutoScaled private constructor(
         val dpToUse = foundEntry?.customValue ?: initialBaseDp
         val finalQualifier = foundEntry?.finalQualifierResolver ?: qualifier
 
-        return dpToUse.toFloat().toDynamicAutoDp(
+        return dpToUse.toDynamicAutoDp(
             context,
             finalQualifier,
             foundEntry?.inverter ?: Inverter.DEFAULT,
