+++ library/src/main/java/com/appdimens/dynamic/code/fill/DimenFillScaled.kt
@@ -290,7 +290,7 @@ class DimenFillScaled private constructor(
         val dpToUse = foundEntry?.customValue ?: initialBaseDp
         val finalQualifier = foundEntry?.finalQualifierResolver ?: qualifier
 
-        return dpToUse.toFloat().toDynamicFillDp(
+        return dpToUse.toDynamicFillDp(
             context,
             finalQualifier,
             foundEntry?.inverter ?: Inverter.DEFAULT,
