+++ library/src/main/java/com/appdimens/dynamic/code/logarithmic/DimenLogarithmicScaled.kt
@@ -290,7 +290,7 @@ class DimenLogarithmicScaled private constructor(
         val dpToUse = foundEntry?.customValue ?: initialBaseDp
         val finalQualifier = foundEntry?.finalQualifierResolver ?: qualifier
 
-        return dpToUse.toFloat().toDynamicLogarithmicDp(
+        return dpToUse.toDynamicLogarithmicDp(
             context,
             finalQualifier,
             foundEntry?.inverter ?: Inverter.DEFAULT,
