+++ library/src/main/java/com/appdimens/dynamic/code/power/DimenPowerScaled.kt
@@ -290,7 +290,7 @@ class DimenPowerScaled private constructor(
         val dpToUse = foundEntry?.customValue ?: initialBaseDp
         val finalQualifier = foundEntry?.finalQualifierResolver ?: qualifier
 
-        return dpToUse.toFloat().toDynamicPowerDp(
+        return dpToUse.toDynamicPowerDp(
             context,
             finalQualifier,
             foundEntry?.inverter ?: Inverter.DEFAULT,
